+++ packages/apptive_grid_user_management/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.1
+* Add option to delete a user's account
+
 ## 0.1.0
 * Promote to 0.1.0 release
 * Update apptive_grid_core to [0.12.0](https://pub.dev/packages/apptive_grid_core/changelog#0120)